Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-dotenv
bcrypt
PyJWT
pandas
openpyxl
python-multipart
email-validator
pymongo
motor
certifi 